+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [2.1.2] - 2021-06-07
+
+- FIX: bug introduced in Vue 2.6.13 (https://github.com/vuejs/vue/issues/12102).
+- WIP: `steinbock` data import.
+
+## [2.1.1] - 2021-06-02
+
+- Per-channel Min-Max / z-Score normalization as a segmentation pre-processing step.
+
 ## [2.1.0] - 2021-05-27
 
 Initial open-source release